requirements.txt
librosa
mcd==0.4
parameter-sherpa==1.0.6
pypesq==1.2.4
pystoi==0.2.2
python-speech-features==0.6
pytorch-ignite==0.3.0
opencv-python==4.1.1.26
scikit-learn==0.22
toolz
tqdm
